fix: rename main class to something different from the main method to avoid confusion.
